requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
firebase-admin>=6.4.0
gunicorn>=21.2.0
firebase-admin